refactor(jwk): fold core JWK shards into facade
