+++ CHANGELOG.md
@@ -26,6 +26,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added perf receipt coverage for seed derivation, JWK/JWKS emission,
   WebAuthn, PKCS#11 mocks, and scanner-safe materialize/verify paths.
 
+### Changed
+
+- Moved JWK/JWKS shape, builder, ordering, and deterministic `kid` internals
+  under `uselesskey-jwk::srp::*`, retaining the former `uselesskey-core-jwk*`,
+  `uselesskey-core-jwks-order`, and `uselesskey-core-kid` crates as
+  published-internal compatibility shims.
+
 ## [0.6.0] - 2026-04-08
 
 ### Added
