+++ crates/uselesskey-core-jwk-builder/Cargo.toml
@@ -5,7 +5,7 @@ edition.workspace = true
 license.workspace = true
 rust-version.workspace = true
 repository.workspace = true
-description = "JWKS builder with deterministic, stable key ordering."
+description = "Deprecated compatibility shim for JWKS builder; prefer uselesskey-jwk."
 categories.workspace = true
 keywords = ["jwk", "jwks", "fixtures", "testing", "deterministic"]
 readme = "README.md"
@@ -15,15 +15,14 @@ documentation = "https://docs.rs/uselesskey-core-jwk-builder"
 authors.workspace = true
 
 [dependencies]
-uselesskey-core-jwk-shape = { path = "../uselesskey-core-jwk-shape", version = "0.6.0" }
-uselesskey-core-jwks-order = { path = "../uselesskey-core-jwks-order", version = "0.6.0" }
-serde_json.workspace = true
+uselesskey-jwk = { path = "../uselesskey-jwk", version = "0.6.0" }
 
 [dev-dependencies]
 insta.workspace = true
 proptest.workspace = true
 serde.workspace = true
 serde_json.workspace = true
+uselesskey-core-jwk-shape = { path = "../uselesskey-core-jwk-shape", version = "0.6.0" }
 
 [package.metadata.docs.rs]
 all-features = true
