+++ crates/uselesskey-core-jwk-shape/Cargo.toml
@@ -5,7 +5,7 @@ edition.workspace = true
 license.workspace = true
 rust-version.workspace = true
 repository.workspace = true
-description = "Core JWK shape types used by JWK/JWKS fixtures."
+description = "Deprecated compatibility shim for JWK/JWKS shapes; prefer uselesskey-jwk."
 categories.workspace = true
 keywords = ["jwk", "jwks", "fixtures", "testing", "jwt"]
 readme = "README.md"
@@ -15,8 +15,7 @@ documentation = "https://docs.rs/uselesskey-core-jwk-shape"
 authors.workspace = true
 
 [dependencies]
-serde.workspace = true
-serde_json.workspace = true
+uselesskey-jwk = { path = "../uselesskey-jwk", version = "0.6.0" }
 
 [dev-dependencies]
 insta.workspace = true
