+++ crates/uselesskey-core-jwks-order/Cargo.toml
@@ -5,7 +5,7 @@ edition.workspace = true
 license.workspace = true
 rust-version.workspace = true
 repository.workspace = true
-description = "Stable kid-based ordering helper for JWKS-like collections."
+description = "Deprecated compatibility shim for JWKS ordering; prefer uselesskey-jwk."
 categories.workspace = true
 keywords = ["jwk", "jwks", "fixtures", "testing", "deterministic"]
 readme = "README.md"
@@ -15,6 +15,7 @@ documentation = "https://docs.rs/uselesskey-core-jwks-order"
 authors.workspace = true
 
 [dependencies]
+uselesskey-jwk = { path = "../uselesskey-jwk", version = "0.6.0" }
 
 [dev-dependencies]
 insta.workspace = true
