+++ crates/uselesskey-core-jwks-order/README.md
@@ -1,8 +1,10 @@
 # uselesskey-core-jwks-order
 
-SRP microcrate for deterministic, insertion-stable sorting of JWK-like items by
-their `kid` value.
+Deprecated compatibility shim for deterministic, insertion-stable sorting of
+JWK-like items by their `kid` value.
 
 - Stable lexicographic ordering by `kid`.
 - Deterministic order for duplicate `kid` values using insertion index.
 - No knowledge of concrete JWK formats.
+
+The implementation now lives in `uselesskey_jwk::srp::ordering`.
