+++ crates/uselesskey-core-keypair-material/Cargo.toml
@@ -16,7 +16,7 @@ authors.workspace = true
 
 [dependencies]
 uselesskey-core = { path = "../uselesskey-core", version = "0.6.0" }
-uselesskey-core-kid.workspace = true
+uselesskey-jwk = { path = "../uselesskey-jwk", version = "0.6.0", default-features = false }
 
 [dev-dependencies]
 insta.workspace = true
