+++ crates/uselesskey-core-kid/Cargo.toml
@@ -5,7 +5,7 @@ edition.workspace = true
 license.workspace = true
 rust-version.workspace = true
 repository.workspace = true
-description = "Deterministic key-id helpers for uselesskey fixture crates."
+description = "Deprecated compatibility shim for deterministic kid helpers; prefer uselesskey-jwk."
 categories.workspace = true
 keywords = ["kid", "jwk", "fixtures", "testing", "deterministic"]
 readme = "README.md"
@@ -15,10 +15,11 @@ documentation = "https://docs.rs/uselesskey-core-kid"
 authors.workspace = true
 
 [dependencies]
-blake3.workspace = true
-base64.workspace = true
+uselesskey-jwk = { path = "../uselesskey-jwk", version = "0.6.0", default-features = false }
 
 [dev-dependencies]
+base64.workspace = true
+blake3.workspace = true
 insta.workspace = true
 proptest.workspace = true
 rstest.workspace = true
