+++ crates/uselesskey-core-kid/README.md
@@ -1,9 +1,10 @@
 # uselesskey-core-kid
 
-Deterministic key-ID (`kid`) helpers for `uselesskey` fixture crates.
+Deprecated compatibility shim for deterministic key-ID (`kid`) helpers now
+owned by `uselesskey-jwk`.
 
-This crate is a small, stable utility layer for turning key bytes into short
-base64url key IDs for tests.
+This crate preserves the previous crate-level API path for published-internal
+consumers. Prefer the public JWK crate for new code.
 
 ## Scope
 
@@ -14,7 +15,7 @@ base64url key IDs for tests.
 ## Example
 
 ```rust
-use uselesskey_core_kid::kid_from_bytes;
+use uselesskey_jwk::srp::kid::kid_from_bytes;
 
 let kid = kid_from_bytes(b"public-key-bytes");
 assert!(!kid.is_empty());
