+++ crates/uselesskey-ed25519/Cargo.toml
@@ -37,7 +37,6 @@ insta.workspace = true
 proptest.workspace = true
 rstest.workspace = true
 serde.workspace = true
-uselesskey-core-kid.workspace = true
 
 
 [lints]
