+++ crates/uselesskey-hmac/Cargo.toml
@@ -16,7 +16,6 @@ authors.workspace = true
 
 [dependencies]
 uselesskey-core = { path = "../uselesskey-core", version = "0.6.0" }
-uselesskey-core-kid.workspace = true
 uselesskey-core-hmac-spec.workspace = true
 rand_chacha10.workspace = true
 rand_core10.workspace = true
