+++ crates/uselesskey-hmac/src/secret.rs
@@ -5,7 +5,7 @@ use rand_chacha10::ChaCha20Rng;
 use rand_core10::{Rng, SeedableRng};
 use uselesskey_core::Factory;
 #[cfg(feature = "jwk")]
-use uselesskey_core_kid::kid_from_bytes;
+use uselesskey_jwk::srp::kid::kid_from_bytes;
 
 use crate::HmacSpec;
