+++ docs/architecture/public-surface.md
@@ -104,7 +104,7 @@ single-responsibility modules under the owning public crate over time.
 | `uselesskey-core-hash` | Published internal | Length-prefixed BLAKE3 derivation helper. | None; derivation is a contract through fixture outputs. | High if exposed because algorithm changes are derivation-sensitive. | `uselesskey_core::srp::hash`. |
 | `uselesskey-core-id` | Published internal | Artifact identity tuple and derivation version inputs. | None; users name labels/specs/variants through public crates. | High if exposed because identity bytes are compatibility-sensitive. | `uselesskey_core::srp::identity`. |
 | `uselesskey-core-seed` | Published internal | Seed parsing and redaction implementation. | Advanced users get `Seed` from `uselesskey-core`. | Medium because `Seed` is public via `uselesskey-core`; crate split is not. | `uselesskey_core::srp::seed`. |
-| `uselesskey-core-kid` | Published internal | Deterministic key-ID generation. | None; users observe `kid()` on fixture outputs. | High if direct imports freeze internals. | `uselesskey_core::srp::kid`. |
+| `uselesskey-core-kid` | Published internal | Compatibility shim for deterministic key-ID generation now owned by `uselesskey-jwk`. | None; users observe `kid()` on fixture outputs. | High if direct imports freeze internals. | `uselesskey_jwk::srp::kid`. |
 | `uselesskey-core-hmac-spec` | Published internal | HMAC spec enum implementation. | Users should import `HmacSpec` from `uselesskey-hmac` or facade. | Medium; spec values are public, crate split is not. | `uselesskey_hmac::srp::spec` or `uselesskey_core::srp::hmac_spec`. |
 | `uselesskey-core-keypair` | Published internal | Shared keypair compatibility facade. | None; key families expose their own fixture handles. | Medium; re-export shell should not become a user dependency. | `uselesskey_core::srp::keypair`. |
 | `uselesskey-core-keypair-material` | Published internal | PKCS#8/SPKI material helpers. | None; users ask fixture crates for PEM/DER outputs. | High if exposed because encoding internals become promises. | `uselesskey_core::srp::keypair_material`. |
@@ -114,10 +114,10 @@ single-responsibility modules under the owning public crate over time.
 | `uselesskey-token-spec` | Published internal | Token spec enum implementation. | Users should import `TokenSpec` from `uselesskey-token` or facade. | Medium; spec values are public, crate split is not. | `uselesskey_token::srp::spec`. |
 | `uselesskey-core-token` | Published internal | Compatibility facade for token-shape primitives. | None; users should import `uselesskey-token`. | Medium; re-export shell should not become a user dependency. | `uselesskey_token::srp::shape`. |
 | `uselesskey-core-token-shape` | Published internal | API key, bearer, JWT-shape, and negative token construction. | None; users should ask `TokenFixture` for values. | High if exposed because scanner-safe shape details become standalone promises. | `uselesskey_token::srp::shape`. |
-| `uselesskey-core-jwk` | Published internal | Core typed JWK/JWKS models behind the facade crate. | Users should import `uselesskey-jwk`. | Medium; JSON shape is public through `uselesskey-jwk`, crate split is not. | `uselesskey_jwk::srp::models`. |
-| `uselesskey-core-jwk-builder` | Published internal | JWKS builder implementation. | Users should import `JwksBuilder` from `uselesskey-jwk`. | Medium; ordering is public, builder crate split is not. | `uselesskey_jwk::srp::builder`. |
-| `uselesskey-core-jwk-shape` | Published internal | Structured JWK/JWKS shapes and negatives. | Users should import typed shapes from `uselesskey-jwk`. | High if exposed because JSON-field internals become standalone promises. | `uselesskey_jwk::srp::shape`. |
-| `uselesskey-core-jwks-order` | Published internal | Stable `kid` ordering helper. | None; users observe stable JWKS output. | Medium; ordering behavior matters, helper crate does not. | `uselesskey_jwk::srp::ordering`. |
+| `uselesskey-core-jwk` | Published internal | Compatibility shim for typed JWK/JWKS models now owned by `uselesskey-jwk`. | Users should import `uselesskey-jwk`. | Medium; JSON shape is public through `uselesskey-jwk`, crate split is not. | `uselesskey_jwk::srp::shape`. |
+| `uselesskey-core-jwk-builder` | Published internal | Compatibility shim for the JWKS builder now owned by `uselesskey-jwk`. | Users should import `JwksBuilder` from `uselesskey-jwk`. | Medium; ordering is public, builder crate split is not. | `uselesskey_jwk::srp::builder`. |
+| `uselesskey-core-jwk-shape` | Published internal | Compatibility shim for structured JWK/JWKS shapes and negatives now owned by `uselesskey-jwk`. | Users should import typed shapes from `uselesskey-jwk`. | High if exposed because JSON-field internals become standalone promises. | `uselesskey_jwk::srp::shape`. |
+| `uselesskey-core-jwks-order` | Published internal | Compatibility shim for stable `kid` ordering now owned by `uselesskey-jwk`. | None; users observe stable JWKS output. | Medium; ordering behavior matters, helper crate does not. | `uselesskey_jwk::srp::ordering`. |
 | `uselesskey-core-x509-spec` | Published internal | X.509 spec models and encoders. | Users should import `X509Spec` / `ChainSpec` from `uselesskey-x509` or facade. | High; cert shape is visible, crate split is not. | `uselesskey_x509::srp::spec`. |
 | `uselesskey-core-x509-derive` | Published internal | X.509 deterministic time, serial, and identity helpers. | None; users observe generated certs. | High if exposed because derivation details become promises. | `uselesskey_x509::srp::derive`. |
 | `uselesskey-core-x509` | Published internal | X.509 policy helper/re-export shell. | Users should import `uselesskey-x509`. | Medium; re-export shell should not become a user dependency. | `uselesskey_x509::srp::policy`. |
