+++ docs/metadata/workspace-docs.json
@@ -263,40 +263,40 @@
       "support_tier": "experimental",
       "publish_status": "published",
       "facade_exposed": false,
-      "semver_expectation": "Internal building blocks; APIs may change between minor releases.",
+      "semver_expectation": "Published-internal compatibility shim; the supported JWK surface is `uselesskey-jwk`.",
       "msrv_policy": "Tracks workspace MSRV (Rust 1.92).",
       "intended_audience": "repo-internal",
-      "deprecation_note": "Prefer depending on `uselesskey` or key-type facade crates unless you are extending the workspace."
+      "deprecation_note": "Prefer `uselesskey-jwk`; this crate is retained only as a compatibility shim."
     },
     {
       "name": "uselesskey-core-jwk-builder",
       "support_tier": "experimental",
       "publish_status": "published",
       "facade_exposed": false,
-      "semver_expectation": "Internal building blocks; APIs may change between minor releases.",
+      "semver_expectation": "Published-internal compatibility shim; the supported JWKS builder surface is `uselesskey-jwk`.",
       "msrv_policy": "Tracks workspace MSRV (Rust 1.92).",
       "intended_audience": "repo-internal",
-      "deprecation_note": "Prefer depending on `uselesskey` or key-type facade crates unless you are extending the workspace."
+      "deprecation_note": "Prefer `uselesskey-jwk::JwksBuilder`; this crate is retained only as a compatibility shim."
     },
     {
       "name": "uselesskey-core-jwk-shape",
       "support_tier": "experimental",
       "publish_status": "published",
       "facade_exposed": false,
-      "semver_expectation": "Internal building blocks; APIs may change between minor releases.",
+      "semver_expectation": "Published-internal compatibility shim; the supported JWK shape surface is `uselesskey-jwk`.",
       "msrv_policy": "Tracks workspace MSRV (Rust 1.92).",
       "intended_audience": "repo-internal",
-      "deprecation_note": "Prefer depending on `uselesskey` or key-type facade crates unless you are extending the workspace."
+      "deprecation_note": "Prefer `uselesskey-jwk`; this crate is retained only as a compatibility shim."
     },
     {
       "name": "uselesskey-core-jwks-order",
       "support_tier": "experimental",
       "publish_status": "published",
       "facade_exposed": false,
-      "semver_expectation": "Internal building blocks; APIs may change between minor releases.",
+      "semver_expectation": "Published-internal compatibility shim; stable JWKS ordering is owned by `uselesskey-jwk`.",
       "msrv_policy": "Tracks workspace MSRV (Rust 1.92).",
       "intended_audience": "repo-internal",
-      "deprecation_note": "Prefer depending on `uselesskey` or key-type facade crates unless you are extending the workspace."
+      "deprecation_note": "Prefer `uselesskey-jwk`; this crate is retained only as a compatibility shim."
     },
     {
       "name": "uselesskey-core-keypair",
@@ -323,10 +323,10 @@
       "support_tier": "experimental",
       "publish_status": "published",
       "facade_exposed": false,
-      "semver_expectation": "Internal building blocks; APIs may change between minor releases.",
+      "semver_expectation": "Published-internal compatibility shim; deterministic kid helpers are owned by `uselesskey-jwk`.",
       "msrv_policy": "Tracks workspace MSRV (Rust 1.92).",
       "intended_audience": "repo-internal",
-      "deprecation_note": "Prefer depending on `uselesskey` or key-type facade crates unless you are extending the workspace."
+      "deprecation_note": "Prefer fixture `kid()` accessors or `uselesskey-jwk`; this crate is retained only as a compatibility shim."
     },
     {
       "name": "uselesskey-core-negative",
