+++ docs/reference/support-matrix.md
@@ -21,13 +21,13 @@
 | `uselesskey-core-hash` | `experimental` | `published` | — | `repo-internal` | Internal building blocks; APIs may change between minor releases. | Tracks workspace MSRV (Rust 1.92). | Prefer depending on `uselesskey` or key-type facade crates unless you are extending the workspace. |
 | `uselesskey-core-hmac-spec` | `experimental` | `published` | — | `repo-internal` | Internal building blocks; APIs may change between minor releases. | Tracks workspace MSRV (Rust 1.92). | Prefer depending on `uselesskey` or key-type facade crates unless you are extending the workspace. |
 | `uselesskey-core-id` | `experimental` | `published` | — | `repo-internal` | Internal building blocks; APIs may change between minor releases. | Tracks workspace MSRV (Rust 1.92). | Prefer depending on `uselesskey` or key-type facade crates unless you are extending the workspace. |
-| `uselesskey-core-jwk` | `experimental` | `published` | — | `repo-internal` | Internal building blocks; APIs may change between minor releases. | Tracks workspace MSRV (Rust 1.92). | Prefer depending on `uselesskey` or key-type facade crates unless you are extending the workspace. |
-| `uselesskey-core-jwk-builder` | `experimental` | `published` | — | `repo-internal` | Internal building blocks; APIs may change between minor releases. | Tracks workspace MSRV (Rust 1.92). | Prefer depending on `uselesskey` or key-type facade crates unless you are extending the workspace. |
-| `uselesskey-core-jwk-shape` | `experimental` | `published` | — | `repo-internal` | Internal building blocks; APIs may change between minor releases. | Tracks workspace MSRV (Rust 1.92). | Prefer depending on `uselesskey` or key-type facade crates unless you are extending the workspace. |
-| `uselesskey-core-jwks-order` | `experimental` | `published` | — | `repo-internal` | Internal building blocks; APIs may change between minor releases. | Tracks workspace MSRV (Rust 1.92). | Prefer depending on `uselesskey` or key-type facade crates unless you are extending the workspace. |
+| `uselesskey-core-jwk` | `experimental` | `published` | — | `repo-internal` | Published-internal compatibility shim; the supported JWK surface is `uselesskey-jwk`. | Tracks workspace MSRV (Rust 1.92). | Prefer `uselesskey-jwk`; this crate is retained only as a compatibility shim. |
+| `uselesskey-core-jwk-builder` | `experimental` | `published` | — | `repo-internal` | Published-internal compatibility shim; the supported JWKS builder surface is `uselesskey-jwk`. | Tracks workspace MSRV (Rust 1.92). | Prefer `uselesskey-jwk::JwksBuilder`; this crate is retained only as a compatibility shim. |
+| `uselesskey-core-jwk-shape` | `experimental` | `published` | — | `repo-internal` | Published-internal compatibility shim; the supported JWK shape surface is `uselesskey-jwk`. | Tracks workspace MSRV (Rust 1.92). | Prefer `uselesskey-jwk`; this crate is retained only as a compatibility shim. |
+| `uselesskey-core-jwks-order` | `experimental` | `published` | — | `repo-internal` | Published-internal compatibility shim; stable JWKS ordering is owned by `uselesskey-jwk`. | Tracks workspace MSRV (Rust 1.92). | Prefer `uselesskey-jwk`; this crate is retained only as a compatibility shim. |
 | `uselesskey-core-keypair` | `experimental` | `published` | — | `repo-internal` | Internal building blocks; APIs may change between minor releases. | Tracks workspace MSRV (Rust 1.92). | Prefer depending on `uselesskey` or key-type facade crates unless you are extending the workspace. |
 | `uselesskey-core-keypair-material` | `experimental` | `published` | — | `repo-internal` | Internal building blocks; APIs may change between minor releases. | Tracks workspace MSRV (Rust 1.92). | Prefer depending on `uselesskey` or key-type facade crates unless you are extending the workspace. |
-| `uselesskey-core-kid` | `experimental` | `published` | — | `repo-internal` | Internal building blocks; APIs may change between minor releases. | Tracks workspace MSRV (Rust 1.92). | Prefer depending on `uselesskey` or key-type facade crates unless you are extending the workspace. |
+| `uselesskey-core-kid` | `experimental` | `published` | — | `repo-internal` | Published-internal compatibility shim; deterministic kid helpers are owned by `uselesskey-jwk`. | Tracks workspace MSRV (Rust 1.92). | Prefer fixture `kid()` accessors or `uselesskey-jwk`; this crate is retained only as a compatibility shim. |
 | `uselesskey-core-negative` | `experimental` | `published` | — | `repo-internal` | Internal building blocks; APIs may change between minor releases. | Tracks workspace MSRV (Rust 1.92). | Prefer depending on `uselesskey` or key-type facade crates unless you are extending the workspace. |
 | `uselesskey-core-negative-der` | `experimental` | `published` | — | `repo-internal` | Internal building blocks; APIs may change between minor releases. | Tracks workspace MSRV (Rust 1.92). | Prefer depending on `uselesskey` or key-type facade crates unless you are extending the workspace. |
 | `uselesskey-core-negative-pem` | `experimental` | `published` | — | `repo-internal` | Internal building blocks; APIs may change between minor releases. | Tracks workspace MSRV (Rust 1.92). | Prefer depending on `uselesskey` or key-type facade crates unless you are extending the workspace. |
